dbms refactoring
Moved some static methods outside of databasemangementsystems class, removed all the for loops that iterated over the database list and table list in favor of a more direct approach of finding a specific table or database i.e. using the list types index method to return the index of the necessary database/table. Also changed the cur_db variable to be a database object instead of a string holding the name of the current db.